feat(ru-client, ru-client-model): write CRUD for cities

+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/CityModel.kt
@@ -2,6 +2,6 @@ package com.fs.client.ru
 
 data class CityModel(
     val id: Int,
-    val countryId: Long,
+    val countryId: Int,
     val name: String
 )
